only unsat and timings

+++ benchmarks/BV-benchamrks-unsat/bvisalpha-64/lisalpha.c
@@ -7,7 +7,7 @@ uint64_t isalpha__cb (uint64_t c) {
 }
 
 
-int foo_cb(int c) {
+uint64_t foo_cb(uint64_t c) {
 	int res = 0;
 	switch (c) {
 	case 0:
@@ -18,8 +18,8 @@ int foo_cb(int c) {
 	return res;
 }
 
-int foo1_cb(int c) {
-	int res = 1;
+uint64_t foo1_cb(uint64_t c) {
+	uint64_t res = 1;
 	switch (c) {
 	case 0:
 	  res = 0;
